chore: refactor phoundry crate
