+++ crates/phoundry/src/lib.rs
@@ -1,2 +1,5 @@
 pub mod error;
-pub mod phorge;
+pub mod test;
+pub mod build;
+pub mod build_and_flatten;
+
